requirements.txt
instagrapi
requests
Pillow
pytesseract
opencv-python
numpy
easyocr
python-dotenv 